LayoutTests/imported/w3c/web-platform-tests/dom/events/document-level-touchmove-event-listener-passive-by-default-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented

Harness Error (FAIL), message = Unhandled rejection: unimplemented

FAIL Touchmove events are non-cancelable since the event listener is treated as passive. promise_test: Unhandled rejection with value: "Condition did not become true after 500 frames"
